+++ packages/common/ProForm/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @harbor-design/proform
 
+## 1.2.6
+
+### Patch Changes
+
+- fix native logic
+
 ## 1.2.5
 
 ### Patch Changes